pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "egile-mcp-client"
version = "0.1.0"
description = "A comprehensive MCP client with direct connection and AI agent modes, including web interface"
authors = [{name = "Jean-Baptiste Poullet", email = "egile@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["mcp", "model-context-protocol", "client", "ai", "agent", "chat", "web"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "httpx>=0.25.0",
    "websockets>=11.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "asyncio-mqtt>=0.11.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "sse-starlette>=1.6.0",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[project.urls]
Homepage = "https://github.com/jpoullet2000/egile-mcp-client"
Repository = "https://github.com/jpoullet2000/egile-mcp-client"

[project.scripts]
egile-mcp-client = "egile_mcp_client.cli:main"

[tool.poetry]
packages = [{include = "egile_mcp_client"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0.0"
httpx = "^0.25.0"
websockets = "^11.0.0"
pydantic = "^2.0.0"
openai = "^1.0.0"
anthropic = "^0.8.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
jinja2 = "^3.1.0"
aiofiles = "^23.0.0"
python-multipart = "^0.0.6"
pyyaml = "^6.0.0"
rich = "^13.0.0"
typer = "^0.9.0"
sse-starlette = "^1.6.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.4.0"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
